文章管理模块:CRUD接口；添加分类、标签删除时，校验是否已关联文章的逻辑。

+++ weblog-module-admin/src/main/java/com/capybara349/weblog/admin/service/impl/AdminTagServiceImpl.java
@@ -1,12 +1,16 @@
 package com.capybara349.weblog.admin.service.impl;
 
+import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.capybara349.weblog.admin.model.vo.tag.*;
 import com.capybara349.weblog.admin.service.AdminTagService;
+import com.capybara349.weblog.common.domain.dos.ArticleTagRelDO;
 import com.capybara349.weblog.common.domain.dos.TagDO;
+import com.capybara349.weblog.common.domain.mapper.ArticleTagRelMapper;
 import com.capybara349.weblog.common.domain.mapper.TagMapper;
 import com.capybara349.weblog.common.enums.ResponseCodeEnum;
+import com.capybara349.weblog.common.exception.BizException;
 import com.capybara349.weblog.common.model.vo.SelectRspVO;
 import com.capybara349.weblog.common.utils.PageResponse;
 import com.capybara349.weblog.common.utils.Response;
@@ -18,6 +22,7 @@ import org.springframework.util.CollectionUtils;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
@@ -31,7 +36,8 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
 
     @Autowired
     TagMapper tagMapper;
-
+    @Autowired
+    private ArticleTagRelMapper articleTagRelMapper;
 
     @Override
     public Response addTag(AddTagReqVO addTagReqVO) {
@@ -87,7 +93,13 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
     public Response deleteTag(DeleteTagReqVO deleteTagReqVO) {
         // 标签 ID
         Long tagId = deleteTagReqVO.getId();
+        // 校验该标签下是否有关联的文章，若有，则不允许删除，提示用户需要先删除标签下的文章
+        ArticleTagRelDO articleTagRelDO = articleTagRelMapper.selectOneByTagId(tagId);
 
+        if (Objects.nonNull(articleTagRelDO)) {
+            log.warn("==> 此标签下包含文章，无法删除，tagId: {}", tagId);
+            throw new BizException(ResponseCodeEnum.TAG_CAN_NOT_DELETE);
+        }
         // 根据标签 ID 删除
         int count = tagMapper.deleteById(tagId);
 
@@ -120,4 +132,24 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
 
         return Response.success(vos);
     }
+
+    @Override
+    public Response findTagSelectList() {
+        // 查询所有标签, Wrappers.emptyWrapper() 表示查询条件为空
+        List<TagDO> tagDOS = tagMapper.selectList(Wrappers.emptyWrapper());
+
+        // DO 转 VO
+        List<SelectRspVO> vos = null;
+        if (!CollectionUtils.isEmpty(tagDOS)) {
+            vos = tagDOS.stream()
+                    .map(tagDO -> SelectRspVO.builder()
+                            .label(tagDO.getName())
+                            .value(tagDO.getId())
+                            .build())
+                    .collect(Collectors.toList());
+        }
+
+        return Response.success(vos);
+    }
+
 }
